+++ scripts/ui/windows/pronoun_creation.py
@@ -9,9 +9,7 @@ from scripts.game_structure.localization import (
 )
 from scripts.cat.pronouns import get_custom_pronouns, add_custom_pronouns
 from scripts.game_structure.screen_settings import MANAGER
-from scripts.game_structure.ui_elements import (
-    UIDropDown,
-)
+from scripts.ui.elements.dropdown import UIDropDown
 from scripts.ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.screens.enums import GameScreen
 from scripts.ui.generate_box import get_box, BoxStyles
